feat(frontend): split auth core and ii config
