+++ src/frontend/src/lib/i18n/zh-cn.json
@@ -364,6 +364,8 @@
 		"main_domain": "主域名（\"派生源\"）",
 		"not_configured": "未配置",
 		"edit_configuration": "使用下面的表单更新您的身份认证配置。",
+		"edit_internet_identity": "在下方配置 Internet Identity 设置。",
+		"edit_google": "在下方配置 Google 身份验证设置。",
 		"main_domain_warn": "您即将更改主域名。请确认这是您期望的操作，因为所有现存用户将不再被您的应用识别。",
 		"external_alternative_origins": "外部备用源",
 		"external_alternative_origins_placeholder": "以逗号分隔的外部域名列表，用于允许派生",
